+++ install/desktop/app-signal.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
+source ~/.local/share/omakub/install/lib/env.sh
 
-wget -qO- https://updates.signal.org/desktop/apt/keys.asc | gpg --dearmor >signal-desktop-keyring.gpg
-cat signal-desktop-keyring.gpg | sudo tee /usr/share/keyrings/signal-desktop-keyring.gpg >/dev/null
-echo 'deb [arch=amd64 signed-by=/usr/share/keyrings/signal-desktop-keyring.gpg] https://updates.signal.org/desktop/apt xenial main' |
-	sudo tee /etc/apt/sources.list.d/signal-xenial.list
-rm signal-desktop-keyring.gpg
-sudo apt update
-sudo apt install -y signal-desktop
+
+if [ "$OS_NAME" = "Ubuntu" ]; then
+	wget -qO- https://updates.signal.org/desktop/apt/keys.asc | gpg --dearmor >signal-desktop-keyring.gpg
+	cat signal-desktop-keyring.gpg | sudo tee /usr/share/keyrings/signal-desktop-keyring.gpg >/dev/null
+	echo 'deb [arch=amd64 signed-by=/usr/share/keyrings/signal-desktop-keyring.gpg] https://updates.signal.org/desktop/apt xenial main' |
+		sudo tee /etc/apt/sources.list.d/signal-xenial.list
+	rm signal-desktop-keyring.gpg
+	sudo $PKG_MGR update
+	sudo $PKG_MGR install -y signal-desktop
+else
+  # Add Signal repository for Fedora
+	flatpak install -y flathub org.signal.Signal
+fi
