+++ install/desktop/optional/app-discord.sh
@@ -1,6 +1,13 @@
+#!/bin/bash
 # A Communication platform for voice, video, and text messaging https://discord.com/
-cd /tmp
-wget https://discord.com/api/download?platform=linux -O discord.deb
-sudo apt install ./discord.deb -y
-rm discord.deb
-cd -
+source ~/.local/share/omakub/install/lib/env.sh
+
+if [ "$OS_NAME" = "Ubuntu" ]; then
+  cd /tmp
+  wget https://discord.com/api/download?platform=linux -O discord.deb
+  sudo $PKG_MGR install ./discord.deb -y
+  rm discord.deb
+  cd -
+else
+  flatpak install flathub com.discordapp.Discord
+fi
